refactor: place prutal alongside other plugins
